ui/requirements.txt
numpy
pandas
openai
pymupdf
pillow
python-dotenv
requests
requests
pydantic
json-repair
tiktoken
tenacity
rich
chainlit
semantic-kernel